+++ xmake.lua
@@ -1,5 +1,5 @@
 set_project("fishnet")
-set_version("2.2.3")
+set_version("2.2.4")
 set_languages("c++20")
 
 option("bedrock")